Add 'creationDate' to Transaction and 'lastTransactionDate' to Account; Make yearMonth non-mandatory in Envelope
Refactor `Transaction` to replace `timestamp` with `creationDate` marked with `@CreatedDate` to ensure automatic population of creation date. Introduce `lastTransactionDate` in `Account` with `@LastModifiedDate` to track the date of the last transaction modification, using the appropriate timezone storage type. Change the `year_month` column in `Envelope` entity to be non-mandatory to allow null values.

Details:
- `Transaction` entity:
  - Removed `timestamp`.
  - Added `creationDate` with `@CreatedDate`.
- `Account` entity:
  - Added `lastTransactionDate` with `@LastModifiedDate` and `@TimeZoneStorageType.AUTO`.
- `Envelope` entity:
  - Made `year_month` column non-mandatory and retained `@JdbcTypeCode(SqlTypes.DATE)`.